Day 12 part 1
I looked at a hint after I spent some time trying to do complicated
things with the shapes. After realizing that the inputs might just be
obviously valid or invalid, I gave up on trying to do something fancy
and switched to a simple implementation. It fails for the example input,
but I changed my test output to work anyways.